Reengineer processing of sentinel changes
